:construction::white_check_mark: Pass first seven tests of bowling challenge, split into seperate class files and refactor code
