+++ part2/bowling/bowling.rb
@@ -1,3 +1,6 @@
+require_relative "turn"
+require_relative "frame"
+
 class Game
   attr_accessor :turn
 
@@ -11,41 +14,5 @@ class Game
 
   def score
     turn.score
-  end
-end
-
-class Turn
-  attr_accessor :frames
-
-  def initialize
-    @frames = [Frame.new]
-  end
-
-  def roll(pins)
-    frames.last << pins
-  end
-
-  def score
-    frames.map(&:score).inject(:+)
-  end
-end
-
-class Frame
-  attr_accessor :goes
-
-  def initialize(pins = nil)
-    @goes = []
-
-    if pins
-      goes << pins
-    end
-  end
-
-  def score
-    goes.inject(0, :+)
-  end
-
-  def <<(pins)
-    goes << pins
   end
 end
