+++ part2/bowling/bowling_test.rb
@@ -32,7 +32,6 @@ class BowlingTest < Minitest::Test
   end
 
   def test_consecutive_spares_each_get_a_one_roll_bonus
-    skip
     game = Game.new
     rolls = [5, 5, 3, 7, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
     rolls.each { |pins| game.roll(pins) }
@@ -40,7 +39,6 @@ class BowlingTest < Minitest::Test
   end
 
   def test_a_spare_in_the_last_frame_gets_a_one_roll_bonus_that_is_counted_once
-    skip
     game = Game.new
     rolls = [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 7, 3, 7]
     rolls.each { |pins| game.roll(pins) }
@@ -48,7 +46,6 @@ class BowlingTest < Minitest::Test
   end
 
   def test_a_strike_earns_ten_points_in_a_frame_with_a_single_roll
-    skip
     game = Game.new
     rolls = [10, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
     rolls.each { |pins| game.roll(pins) }
